requirements.txt
Pillow==11.2.1